api/requirements.txt
Flask>=2.2.0,<3.0.0
python-dotenv>=0.20.0,<1.0.0
Flask-Cors>=3.0.0,<4.0.0
PyYAML>=5.4,<7.0
Jinja2>=3.0,<4.0
openai>=1.0.0,<2.0.0
redis>=4.0.0,<6.0.0
supabase>=1.0.0,<2.0.0
# LLM Client specific dependencies
google-generativeai>=0.3.0,<0.6.0
anthropic>=0.20.0,<0.30.0
groq>=0.5.0,<0.9.0
huggingface_hub>=0.20.0,<0.24.0
requests>=2.20.0,<3.0.0

